changed dtype of molnums from int to np.int64

+++ package/MDAnalysis/core/topologyattrs.py
@@ -1460,7 +1460,7 @@ class Molnums(ResidueAttr):
     attrname = 'molnums'
     singular = 'molnum'
     target_classes = [AtomGroup, ResidueGroup, Atom, Residue]
-    dtype = int
+    dtype = np.int64
 
 # segment attributes
